+++ 204/9.cpp
@@ -6,12 +6,12 @@
 
 int main()
 {
-    std::vector<float> First;
+    std::vector<double> First;
     std::string str;
     std::getline(std::cin, str);
     std::stringstream inp;
     inp << str;
-    float a = -1; 
+    double a = -1; 
     size_t cnt = 0;
     for(float tmp = 0; inp >> tmp;){
         if (cnt == 1)
@@ -20,7 +20,7 @@ int main()
             First.push_back(tmp);
         ++cnt;
     }
-    std::vector<float> Ans(First.size() + 1, 0);
+    std::vector<double> Ans(First.size() + 1, 0);
     
     for (long long int i = 0; i < First.size(); ++i){
         Ans[i + 1] += First[i];
